RegionContainerIjRoi: bug fix for 3D objects with empty slices. Region: added isEmpty method TiledDiskBackedImage: bugfix getZPlane BinaryOperation: bugfix on matching when object bounding boxes overlap Selection : added validation prompt on delete

+++ bacmman-gui/src/main/java/bacmman/ui/gui/selection/SelectionUtils.java
@@ -707,6 +707,7 @@ public class SelectionUtils {
         JMenuItem delete = new JMenuItem("Delete Selection");
         delete.addActionListener((ActionEvent e) -> {
             if (selectedValues.isEmpty()) return;
+            if (!Utils.promptBoolean("Permanently delete selected selection (s) ?", list)) return;
             DefaultListModel<Selection> model = (DefaultListModel<Selection>)list.getModel();
             for (Selection s : selectedValues ) dao.delete(s);
             for (Selection s : selectedValues) model.removeElement(s);
